remove pygame keyboard input
All the local keyboard input has been removed, since this used pygame which was interfering with the systemd startup stuff. This version can only be controlled via MQTT messages sent from the cyberdeck, so the hand needs to be connected to cyberdeck via ethernet cable, instead of using a USB keyboard.